refactor: rename 'metrics' -> 'geometry'
