+++ src/geometry/get-size.ts
