+++ src/geometry/is-visible.ts
